requirements.txt
# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# LLM Integration
openai>=1.3.0
google-generativeai>=0.3.0  # Add this for Gemini
# anthropic>=0.7.0  # Remove or comment out
langchain>=0.1.0

# Core data analysis libraries
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Data visualization
plotly>=5.17.0

# Data Sources & Databases
requests>=2.31.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# File handling
aiofiles>=23.2.0
openpyxl>=3.1.0

# Web scraping and HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0 
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For async testing 